requirements.txt
aiosignal==1.2.0
anyio==3.6.2
astor==0.8.1
async-timeout==4.0.2
attrs==22.1.0
bandit==1.7.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
darglint==1.8.1
docutils==0.19
eradicate==2.1.0
fastapi==0.86.0
flake8==4.0.1
flake8-bandit==3.0.0
flake8-broken-line==0.5.0
flake8-bugbear==22.10.27
flake8-commas==2.1.0
flake8-comprehensions==3.10.1
flake8-debugger==4.1.2
flake8-docstrings==1.6.0
flake8-eradicate==1.4.0
flake8-isort==4.2.0
flake8-polyfill==1.0.2
flake8-quotes==3.3.1
flake8-rst-docstrings==0.2.7
flake8-string-format==0.3.0
frozenlist==1.3.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.29
h11==0.14.0
idna==3.4
isort==5.10.1
Jinja2==3.1.2
loguru==0.6.0
MarkupSafe==2.1.1
mccabe==0.6.1
multidict==6.0.2
mypy==0.982
mypy-extensions==0.4.3
orjson==3.8.1
pbr==5.11.0
pep8-naming==0.13.2
pocketsphinx==5.0.0
PyAudio==0.2.12
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.2
pydocstyle==6.1.1
pydub==0.25.1
pyflakes==2.4.0
Pygments==2.13.0
PyJWT==2.6.0
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
restructuredtext-lint==1.4.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sounddevice==0.4.5
SpeechRecognition==3.8.1
srt==3.5.2
starlette==0.20.4
stevedore==4.1.0
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
vosk==0.3.44
websockets==10.4
wemake-python-styleguide==0.17.0
yarl==1.8.1
aiohttp~=3.8.3
numpy==1.23.4
pytest==7.2.0
pytest-asyncio==0.20.2
spacy~=3.4.3

aioredis~=2.0.1